ndtest: Make ndtest a module on its own
Today ndtest module is compiled as nfit_test.ko depending on
if the CONFIG_ACPI_NFIT defined or not.

It is more advantageous to make ndtest a module on its own
so that the unit tests can be run serially on the same host
without a need for recompilation of sources.

The patch modifies the Kbuild file to takes care of that.

Signed-off-by: Shivaprasad G Bhat <sbhat@linux.ibm.com>